requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
pandas
scikit-learn
streamlit
mlxtend
